clean up

+++ zeta/utils/attention/flash_triton.py
@@ -74,7 +74,9 @@ def _fwd_kernel(
         qk = tl.zeros([BLOCK_M, BLOCK_N], dtype=tl.float32)
         if IS_CAUSAL:
             qk = tl.where(offs_m[:, None] >= (start_n + offs_n[None, :]), qk, float("-inf"))
-        qk += tl.dot(q, k)
+        # qk += tl.dot(q, k)
+        qk += tl.dot(q.to(tl.float32), k)  # convert q to fp32
+        ###################
         # -- compute scaling constant ---
         m_i_new = tl.maximum(m_i, tl.max(qk, 1))
         alpha = tl.math.exp2(m_i - m_i_new)
